requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
tenacity==8.2.3